loadbalancer: optimizations to idAllocator
[ upstream commit c4a83dbd0553cd2e7abc03181557757750751c41 ]

- Store L3n4AddrID by value in idAllocator to reduce allocations
- Use L3n4Addr as the entities map key
